ISL_testing/requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
torch>=2.0.0
transformers>=4.40.0
pillow>=10.0.0
